Reorganizes CSS settings into separate files (per library) -- CW-715
