+++ src/features/ToolBar/DropdownMenu.tsx
@@ -1,5 +1,3 @@
-import './DataMenu/menuItem.css'
-
 import { Button, Tooltip } from '@mui/material'
 import { Box } from '@mui/material'
 import { useTheme } from '@mui/material/styles'
