+++ src/styles/index.css
@@ -1,0 +1,5 @@
+@import "./base.css";
+@import "./allotment.css";
+@import "./glide.css";
+@import "./mantine.css";
+@import "./primereact.css";
