+++ src/styles/mantine.css
@@ -1,0 +1,35 @@
+/**
+ * This file contains styles for the Mantine components.
+ * 
+ * IMPORTANT: These styles are meant to override the default Mantine styles 
+ *            to better match the MUI theme used in the application.
+ */
+
+/** Input */
+
+.mantine-Input-input {
+  background-color: var(--mui-palette-background-paper);
+  color: var(--mui-palette-text-primary);
+  border-color: var(--mui-palette-divider);
+}
+.mantine-Input-input:hover {
+  border-color: var(--mui-palette-text-primary);
+}
+.mantine-Input-input:focus {
+  border-color: var(--mui-palette-primary-main);
+}
+
+/** Number Input Controls (up/down buttons) */
+
+.mantine-NumberInput-control {
+  border-inline-start: 1px solid var(--mui-palette-divider);
+}
+.mantine-NumberInput-control:hover {
+  background-color: var(--mui-palette-action-hover);
+}
+.mantine-NumberInput-control svg {
+  color: var(--mui-palette-text-secondary);
+}
+.mantine-NumberInput-control:hover svg {
+  color: var(--mui-palette-text-primary);
+}
